+++ config/app.php
@@ -1,6 +1,7 @@
 <?php
+
 return [
-    'name' => env('APP_NAME', 'L2CMS'),
+    'name' => env('APP_NAME', 'L2Forge CMS'),
     'env' => env('APP_ENV', 'production'),
     'debug' => (bool) env('APP_DEBUG', false),
     'url' => env('APP_URL', 'http://localhost'),