+++ config/cache.php
@@ -7,5 +7,5 @@ return [
         'file' => ['driver' => 'file', 'path' => storage_path('framework/cache/data'), 'lock_path' => storage_path('framework/cache/data')],
         'redis' => ['driver' => 'redis', 'connection' => env('REDIS_CACHE_CONNECTION', 'cache'), 'lock_connection' => env('REDIS_CACHE_LOCK_CONNECTION', 'default')],
     ],
-    'prefix' => env('CACHE_PREFIX', 'l2cms_cache_'),
+    'prefix' => env('CACHE_PREFIX', 'l2forge_cache_'),
 ];